Add files via upload
The code is used for predicting if the customers will return the clothes or not. It include data processing, features selection, EDA and model section. I choose logistic regression and decision tree as my main models. Meanwhile, I used the cost senstive method to find the suitable cutoff with minimum wrongly prediction costs. Finally, the AUC decreases but the cost  is the minimum